+++ CHANGES.md
@@ -9,3 +9,4 @@
 - Fix update user profile
 - Add Address validation < 10 kms
 - Check user phone exist before placing a new order
+- Add minimum purchase charge